// frontend/src/openapi/requests/types.gen.ts
export type CompanySearchFilters = {
  industries?: Array<string> | null;
  subIndustries?: Array<string> | null;
  numericRanges?: {
    [key: string]: NumericRange;
  } | null;
};

export type CompanySearchResultSchema = {
  id: string;
  name: string;
  ticker: string;
  industry?: string | null;
  equity_value?: number | null;
  ltm_revenue?: number | null;
  multiple_ev_to_revenue?: number | null;
  created_at: string;
  updated_at: string;
};

export type CompanySearchSchema = {
  search?: string | null;
  filters?: CompanySearchFilters | null;
  sorting?: Array<SortCriterion> | null;
  pagination?: Pagination | null;
};
// ...
export type type = "10-Q" | "10-K";

export type source = "EDGAR" | "Manual" | "Other";

export type NumericRange = {
  min?: number | null;
  max?: number | null;
};

export type Pagination = {
  offset?: number | null;
  limit?: number | null;
};

export type SortCriterion = {
  field: string;
  direction: "asc" | "desc";
};

export type direction = "asc" | "desc";